Simplifying docstring for ``views.say``.
I decided to quit copying all of the official documentation into the
docstrings. It takes up way too much space, and is really annoying for me to
format by hand. Instead--I'll use my own sane descriptions for the views, and
just link to the official docs.

I think this is a better long-term decision.