Re-building documentation structure.
Using the separate build / source directories for our Sphinx docs, since as the
documentation is growing, it's becoming uglier and uglier to have it all in one
place.

Eventually I'd like to move to a heiarchal documentation sorta scheme, so we can
group various chapters (if you will) into their own subfolders.