Fixing broken ``test_blacklist_works`` test case.
It was broken because I had hard-coded in the XML response text to check
against--and during the lastest twilio helper library update, the XML generator
changed a bit.

(Specifically, utf-8 => UTF-8.) But anyhow, this test case is now more
future-proof, as it actually uses the twiml library to verify the rejection.